requirements.txt
# project requirements
pyaudio
openai-whisper
pydub
scipy

# issues related to whisper, particularly ffmpeg
# https://github.com/openai/whisper